requirements.txt
Django>=5.2.3
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
celery>=5.3.0
djangorestframework
pytesseract
paddleocr
paddlepaddle
psutil>=5.9.0